New identity provider, improved build script, etc
New identity provider: Manual Identity Provider allows you to manually enter a MAC address to connect to.

Build script now extended to allow for copying necessary files to a project using the library.
To do this, you must enter the path to the project into the local.properties file in the library folder: Make a line containing 'mainproject=<path>'
Now the command 'ant updatetest' will copy the library JAR-file as well as the relevant resources into the project folder. Please note that it is necessary to refresh the libs-folder of the project in Eclipse whenever you run 'ant updatetest' to ensure that the newest version of the JAR gets included with the Android app.

Resources needed from within the library earlier was reached using
  R.layout.main (or something like that)
Now they must be reached using
  BtAPI.res( <Context>, "layout", "main" )
This was necessary to use resources from the library in an application.