A basic test project have been set up under the API project. This is done for it to be able to test the library's functionality.
In order to run the tests, the following commands need to be run from a terminal, under the test directory:
- rm -rf bin/* gen/*
- ant install
- adb shell am instrument -w dk.hotmovinglobster.dustytuba.tests/android.test.InstrumentationTestRunner

The first might not be necesary, but I had some trouble with the already compiled code.
The second (re)installs the library on a device or emulator.
The third starts the tests.

The installment is required whenever you write new tests.